Práctica: fichero de configuración + reorg de la pantalla + BPM con decimal
Feedback del autor. Cierra F.13 (ADR-063, ADR-064).

Fichero de configuración (ADR-063)
- SettingsStore: AppSettings -> JSON legible/copiable en
  ~/Library/Application Support/xFlare/settings.json, escrito de forma
  atómica en cada cambio. Antes solo UserDefaults, que cfprefsd vaciaba
  tarde y perdía cambios si la app se cerraba de golpe ("las canciones y
  la configuración no se guardan de una vez a otra"). loadSettings
  prefiere el fichero y migra el plist viejo una vez.

Reorg de la pantalla de práctica (ADR-064)
- Columna izquierda (leftColumn): todo lo del SAMPLE — selector, 4 slots
  MIDI (el número dispara loadSlot en caliente, el menú asigna un
  fichero), cue A/B, meter, EQ Lo/Mid/Hi, volúmenes.
- Zona inferior (bottomBar): fila compacta con el nombre de la base y, a
  la derecha, tempo + rejilla (reiniciar, ÷2/×2, ◀/▶, TAP, BPM). El botón
  ▴ despliega la lista de instrumentales ya analizadas de la Librería; al
  cargar una base se minimiza sola.
- Panel derecho: solo "Repite conmigo", "Grabar línea", "Ajuste rápido".
- LivePracticeView gana onSampleSlotsChanged (cableado en AppRootView).

BPM con un decimal
- PracticeSession.bpm Int -> Double (setBPM redondea a 0,1).
- TapTempo.tap() -> Double? (media de 4-8 golpes, a 0,1).
- PlatterInputView.onBPM / currentBPM -> Double.
- La UI formatea con %.1f; la edición a mano acepta coma o punto.

Tope del sample de scratch
- AudioAsset.scratchMaxSeconds (2 s) + capScratch: un fichero largo ya no
  barre minutos de audio con un gesto ("se va todo"); cualquier sample se
  scratchea como el Ahh y el rail muestra los mismos segundos.

Números de la rejilla
- PracticeScene.gridLabels (puro): "compás.subdivisión" (1.1, 1.2…) sobre
  cada línea de negra, arriba, pequeños y discretos; siguen a gridShift.

638 tests en verde (make verify). Docs: ADR-063/064, UI_DESIGN 3.3b/3.7,
MODULE_STATUS, TODO F.13, backlog.json, README.

+++ Sources/XFApp/PlatterInputView.swift
@@ -24,8 +24,8 @@ struct PlatterInputView: NSViewRepresentable {
     var onFreeze: () -> Void
     var onCue: () -> Void
     var onRestartInstrumental: () -> Void = {}
-    var onBPM: (Int) -> Void
-    var currentBPM: () -> Int
+    var onBPM: (Double) -> Void
+    var currentBPM: () -> Double
     var onExit: () -> Void
 
     func makeNSView(context: Context) -> NSView { CatcherView(owner: self) }
